If all else fails choose a random valid move
