+++ game/game.go
@@ -1,20 +1,85 @@
 package game
 
+import(
+	"math/rand"
+	"time"
+	"github.com/SamuelWillis/go_snake/helpers"
+)
+
 // GetNextMove that our snake will do.
 func GetNextMove(state State) string {
 	var move string
+	validMoves := getValidMoves(state)
 
 	if shouldEatFood() {
-		move = getMoveToFood(state)
+		move = getMoveToFood(state, validMoves)
 	}
 
 	if shouldAttack() {
-		move = getMoveToFood(state)
+		move = getMoveToFood(state, validMoves)
 	}
 
 	if move == "" {
-		move = "left"
+		move = chooseRandomMove(validMoves)
 	}
 
 	return move
+}
+
+func getValidMoves(state State) ValidMoves {
+	head := state.You.Body[0]
+	snakes := state.Board.Snakes
+
+	validMoves := ValidMoves{
+		up: isValidMove(Coord{ X: head.X, Y: head.Y - 1 }, snakes),
+		down: isValidMove(Coord{ X: head.X, Y: head.Y + 1 }, snakes),
+		left: isValidMove(Coord{ X: head.X - 1, Y: head.Y }, snakes),
+		right: isValidMove(Coord{ X: head.X + 1, Y: head.Y }, snakes),
+	}
+
+	helpers.Dump("up", validMoves.up)
+	helpers.Dump("down", validMoves.down)
+	helpers.Dump("left", validMoves.left)
+	helpers.Dump("right", validMoves.right)
+
+	return validMoves
+}
+
+func isValidMove(test Coord, snakes []Snake) bool {
+	isValid := true
+	helpers.Dump("testing", test)
+	helpers.Dump("", "")
+	for i := 0; i < len(snakes); i++ {
+		body := snakes[i].Body
+		for j := 0; j < len(body); j++ {
+			helpers.Dump("body", body[j])
+			if (test == body[j]) {
+				isValid = false
+			}
+		}
+	}
+	return isValid
+}
+
+func chooseRandomMove(validMoves ValidMoves) string {
+	rand.Seed(time.Now().Unix())
+	var moves []string
+
+	if validMoves.up {
+		moves = append(moves, "up")
+	}
+
+	if validMoves.down {
+		moves = append(moves, "down")
+	}
+
+	if validMoves.left {
+		moves = append(moves, "left")
+	}
+
+	if validMoves.right {
+		moves = append(moves, "right")
+	}
+
+	return moves[rand.Intn(len(moves))]
 }
